Max Area of Island
Medium Leet Code problem no. 695 Max Area of Island: https://leetcode.com/problems/max-area-of-island/, basic BFS solution; if grid[i][j] equals to 1 we add all 4 connected fields to the stack, and while removing them we check if they have valid coordinates and are equal to one - if so we add +1 to our actual score and repeat the process.